igb: Fix error handlings with realization
For msix_vector_use and pcie_endpoint_cap_init, the error is tested by
assert() as it should never report an error if the implementation is
correct, and continuing the execution can result in an unexpected
situation.

For msix_init, it is handled by canceling the device realization. The
cancellation will be detected by looking at the output of errp
parameter.

This change also adds a check for igbvf if it is created as a virtual
function. It can be otherwise if a user specified igbvf with -device
command line argument.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>