virtio-net: Enable software RSS
virtio-net implements software RSS but does not enable it. Enable it
when RSS is requested, but the eBPF implementation is not available.
We also check if vhost is in use in such a case since software RSS is
incompatible with vhost. A warning will be emitted when falling back to
software RSS since it provides no performance benefit.

Signed-off-by: Akihiko Odaki <akihiko.odaki@daynix.com>